Conflicts:
	language/English/strings.po
	project/VS2010Express/XBMC.vcxproj.filters
	xbmc/utils/MarkWatchedJob.cpp
	xbmc/utils/TimeUtils.h
	xbmc/video/dialogs/GUIDialogVideoSettings.cpp

+++ xbmc/addons/AddonCallbacksAddon.h
@@ -47,9 +47,9 @@ public:
   // file operations
   static void* OpenFile(const void* addonData, const char* strFileName, unsigned int flags);
   static void* OpenFileForWrite(const void* addonData, const char* strFileName, bool bOverwrite);
-  static unsigned int ReadFile(const void* addonData, void* file, void* lpBuf, int64_t uiBufSize);
+  static ssize_t ReadFile(const void* addonData, void* file, void* lpBuf, size_t uiBufSize);
   static bool ReadFileString(const void* addonData, void* file, char *szLine, int iLineLength);
-  static int WriteFile(const void* addonData, void* file, const void* lpBuf, int64_t uiBufSize);
+  static ssize_t WriteFile(const void* addonData, void* file, const void* lpBuf, size_t uiBufSize);
   static void FlushFile(const void* addonData, void* file);
   static int64_t SeekFile(const void* addonData, void* file, int64_t iFilePosition, int iWhence);
   static int TruncateFile(const void* addonData, void* file, int64_t iSize);
