+++ xbmc/filesystem/VTPFile.h
@@ -41,7 +41,7 @@ public:
   virtual int64_t       GetLength()                                    { return -1; }
   virtual int           Stat(const CURL& url, struct __stat64* buffer) { return -1; }
   virtual void          Close();
-  virtual unsigned int  Read(void* buffer, int64_t size);
+  virtual ssize_t       Read(void* buffer, size_t size);
   virtual std::string   GetContent()                                   { return ""; }
   virtual bool          SkipNext()                                     { return m_socket ? true : false; }
